Clean code for generate manuscript data [SCI-6663]
