+++ app/controllers/api/service/base_controller.rb
@@ -53,6 +53,13 @@ module Api
         )
       end
 
+      rescue_from URI::InvalidURIError do |e|
+        render_error(
+          I18n.t('api.service.errors.url.not_valid'), 
+          I18n.t('api.service.errors.url.not_valid'), :bad_request
+        )
+      end
+
       rescue_from JWT::DecodeError,
                   JWT::InvalidPayload,
                   JWT::VerificationError,
