+++ config/locales/en.yml
@@ -3001,6 +3001,8 @@ en:
       errors:
         missing_data: "Missing data"
         callback_missing: "Missing callback"
+        url:
+          not_valid: "Not valid URL"
 
 
   Add: "Add"
